Got rid of /helpers folder in favor of simply using folders with the appropriate component names; slightly refactored lib/ApiRequest structure so that internal types are not exported.
