+++ app/components/PageWrapper/NavigationBar/Logo/index.js
