+++ app/forms/UserForm/index.js
@@ -8,7 +8,7 @@ import { type FormProps } from 'redux-form/lib/types.js.flow';
 import { translate, type TranslatorProps } from 'react-i18next';
 import { Form, Input, Checkbox } from 'semantic-ui-react';
 
-import SubmitButtonGroup from '../helpers/SubmitButtonGroup';
+import SubmitButtonGroup from 'components/SubmitButtonGroup';
 
 type UserFormValues = {|
   email?: string,
