+++ app/lib/ApiRequest/ApiRequest/index.test.js
@@ -4,11 +4,13 @@
 import { API_URL } from 'config/api';
 import { UnsupportedOperationError } from 'errors';
 
-import fetchApiResponseData from './helpers/fetchApiResponseData';
-import ApiRequest, { defaultConfig } from './ApiRequest';
-import * as m from './model';
+import { httpMethods } from '../types';
 
-jest.mock('./helpers/fetchApiResponseData');
+import fetchApiResponseData from './fetchApiResponseData';
+
+import ApiRequest, { defaultConfig } from '.';
+
+jest.mock('./fetchApiResponseData');
 
 describe(`ApiRequest`, (): void => {
 
@@ -22,22 +24,22 @@ describe(`ApiRequest`, (): void => {
   });
 
   it(`correctly sets the method and uses defaults for the rest of the config, when it is created`, (): void => {
-    const dummyRequest = new ApiRequest(m.httpMethods.GET);
+    const dummyRequest = new ApiRequest(httpMethods.GET);
     expect(dummyRequest.config).toEqual({
       ...defaultConfig,
-      method: m.httpMethods.GET,
+      method: httpMethods.GET,
     });
   });
 
   it(`calls fetchApiResponseData with the correct arguments, when it is executed`, (): void => {
-    new ApiRequest(m.httpMethods.GET)
+    new ApiRequest(httpMethods.GET)
       .addPathSegment('test')
       .execute();
 
     expect(dummyFetchApiResponseData).toHaveBeenCalledWith(
       `${API_URL}/test`,
       {
-        method: m.httpMethods.GET,
+        method: httpMethods.GET,
         headers: defaultConfig.headers,
         body: null,
       },
@@ -45,14 +47,14 @@ describe(`ApiRequest`, (): void => {
   });
 
   it(`strips a starting slash from a pathSegment, when a pathSegment with a starting slash is added`, (): void => {
-    new ApiRequest(m.httpMethods.GET)
+    new ApiRequest(httpMethods.GET)
       .addPathSegment('/test')
       .execute();
 
     expect(dummyFetchApiResponseData).toHaveBeenCalledWith(
       `${API_URL}/test`,
       {
-        method: m.httpMethods.GET,
+        method: httpMethods.GET,
         headers: defaultConfig.headers,
         body: null,
       },
@@ -60,7 +62,7 @@ describe(`ApiRequest`, (): void => {
   });
 
   it(`executes with extra parameters, when extra parameters are passed before execution`, (): void => {
-    new ApiRequest(m.httpMethods.GET)
+    new ApiRequest(httpMethods.GET)
       .addPathSegment('test')
       .setParameter('page', '5')
       .setParameter('filter', 'foobar')
@@ -69,7 +71,7 @@ describe(`ApiRequest`, (): void => {
     expect(dummyFetchApiResponseData).toHaveBeenCalledWith(
       `${API_URL}/test?page=5&filter=foobar`,
       {
-        method: m.httpMethods.GET,
+        method: httpMethods.GET,
         headers: defaultConfig.headers,
         body: null,
       },
@@ -77,7 +79,7 @@ describe(`ApiRequest`, (): void => {
   });
 
   it(`allows previously set parameters to be overridden with new values`, (): void => {
-    new ApiRequest(m.httpMethods.GET)
+    new ApiRequest(httpMethods.GET)
       .addPathSegment('test')
       .setParameter('page', '5')
       .setParameter('page', '8')
@@ -86,7 +88,7 @@ describe(`ApiRequest`, (): void => {
     expect(dummyFetchApiResponseData).toHaveBeenCalledWith(
       `${API_URL}/test?page=8`,
       {
-        method: m.httpMethods.GET,
+        method: httpMethods.GET,
         headers: defaultConfig.headers,
         body: null,
       },
@@ -94,7 +96,7 @@ describe(`ApiRequest`, (): void => {
   });
 
   it(`executes with the correct url, when multiple pathSegments and parameters are passed`, (): void => {
-    new ApiRequest(m.httpMethods.GET)
+    new ApiRequest(httpMethods.GET)
       .addPathSegment('lorem')
       .addPathSegment('ipsum')
       .setParameter('page', '5')
@@ -104,7 +106,7 @@ describe(`ApiRequest`, (): void => {
     expect(dummyFetchApiResponseData).toHaveBeenCalledWith(
       `${API_URL}/lorem/ipsum?page=5&filter=foobar`,
       {
-        method: m.httpMethods.GET,
+        method: httpMethods.GET,
         headers: defaultConfig.headers,
         body: null,
       },
@@ -113,13 +115,13 @@ describe(`ApiRequest`, (): void => {
 
   it(`throws an UnsupportedOperationError, when attempting to execute without having added any pathSegments`, (): void => {
     expect((): void => {
-      new ApiRequest(m.httpMethods.GET)
+      new ApiRequest(httpMethods.GET)
         .execute();
     }).toThrow(UnsupportedOperationError);
   });
 
   it(`executes with extra headers, when extra headers are passed before execution`, (): void => {
-    new ApiRequest(m.httpMethods.GET)
+    new ApiRequest(httpMethods.GET)
       .addPathSegment('test')
       .setHeader('Host', 'localhost')
       .setHeader('Accept-Charset', 'utf-8')
@@ -128,7 +130,7 @@ describe(`ApiRequest`, (): void => {
     expect(dummyFetchApiResponseData).toHaveBeenCalledWith(
       `${API_URL}/test`,
       {
-        method: m.httpMethods.GET,
+        method: httpMethods.GET,
         headers: {
           ...defaultConfig.headers,
           'Host': 'localhost',
@@ -140,7 +142,7 @@ describe(`ApiRequest`, (): void => {
   });
 
   it(`allows default headers to be overridden with new values`, (): void => {
-    new ApiRequest(m.httpMethods.GET)
+    new ApiRequest(httpMethods.GET)
       .addPathSegment('test')
       .setHeader('Content-Type', 'text/html')
       .execute();
@@ -148,7 +150,7 @@ describe(`ApiRequest`, (): void => {
     expect(dummyFetchApiResponseData).toHaveBeenCalledWith(
       `${API_URL}/test`,
       {
-        method: m.httpMethods.GET,
+        method: httpMethods.GET,
         headers: {
           ...defaultConfig.headers,
           'Content-Type': 'text/html',
@@ -160,7 +162,7 @@ describe(`ApiRequest`, (): void => {
 
   it(`executes with the correct body, when a body is passed before execution`, (): void => {
     const dummyBody = JSON.stringify({ data: { foo: 'bar' } });
-    new ApiRequest(m.httpMethods.POST)
+    new ApiRequest(httpMethods.POST)
       .addPathSegment('test')
       .setBody(dummyBody)
       .execute();
@@ -168,7 +170,7 @@ describe(`ApiRequest`, (): void => {
     expect(dummyFetchApiResponseData).toHaveBeenCalledWith(
       `${API_URL}/test`,
       {
-        method: m.httpMethods.POST,
+        method: httpMethods.POST,
         headers: defaultConfig.headers,
         body: dummyBody,
       },
@@ -179,14 +181,14 @@ describe(`ApiRequest`, (): void => {
     const dummyBody = JSON.stringify({ data: { foo: 'bar' } });
 
     expect((): void => {
-      new ApiRequest(m.httpMethods.GET)
+      new ApiRequest(httpMethods.GET)
         .addPathSegment('test')
         .setBody(dummyBody);
     }).toThrow(UnsupportedOperationError);
   });
 
   it(`executes with the correct Auth header, when a token is passed before execution`, (): void => {
-    new ApiRequest(m.httpMethods.POST)
+    new ApiRequest(httpMethods.POST)
       .addPathSegment('test')
       .setToken('foobarToken')
       .execute();
@@ -194,7 +196,7 @@ describe(`ApiRequest`, (): void => {
     expect(dummyFetchApiResponseData).toHaveBeenCalledWith(
       `${API_URL}/test`,
       {
-        method: m.httpMethods.POST,
+        method: httpMethods.POST,
         headers: {
           ...defaultConfig.headers,
           'Authorization': 'Bearer foobarToken',
@@ -205,7 +207,7 @@ describe(`ApiRequest`, (): void => {
   });
 
   it(`allows unsetting the Auth header by passing null as a token`, (): void => {
-    new ApiRequest(m.httpMethods.POST)
+    new ApiRequest(httpMethods.POST)
       .addPathSegment('test')
       .setToken('foobarToken')
       .setToken(null)
@@ -214,7 +216,7 @@ describe(`ApiRequest`, (): void => {
     expect(dummyFetchApiResponseData).toHaveBeenCalledWith(
       `${API_URL}/test`,
       {
-        method: m.httpMethods.POST,
+        method: httpMethods.POST,
         headers: defaultConfig.headers,
         body: null,
       },
@@ -222,7 +224,7 @@ describe(`ApiRequest`, (): void => {
   });
 
   it(`allows unsetting the Auth header by passing an empty string as a token`, (): void => {
-    new ApiRequest(m.httpMethods.POST)
+    new ApiRequest(httpMethods.POST)
       .addPathSegment('test')
       .setToken('foobarToken')
       .setToken('')
@@ -231,7 +233,7 @@ describe(`ApiRequest`, (): void => {
     expect(dummyFetchApiResponseData).toHaveBeenCalledWith(
       `${API_URL}/test`,
       {
-        method: m.httpMethods.POST,
+        method: httpMethods.POST,
         headers: defaultConfig.headers,
         body: null,
       },
