+++ app/modules/platform/components/Sidebars/sidebarIdsToComponentsMap/Sidebar/index.js
