+++ app/modules/platform/components/Sidebars/sidebarIdsToComponentsMap/SlidePreviewsSidebar.js
@@ -6,7 +6,7 @@ import { translate, type TranslatorProps } from 'react-i18next';
 import Slides from 'components/Slides';
 import topics from 'modules/topics';
 
-import Sidebar from './helpers/Sidebar';
+import Sidebar from './Sidebar';
 
 type PassedProps = {|
   topic: topics.model.Topic,
