+++ app/modules/platform/components/SidebarsMenu/SidebarsMenuItem/index.js
